extract tunnel plugin types to its own file
